replaced the teachapi mixin with a Higher Order Component, and updated all dependents and tests accordingly
