+++ pages/clubs.jsx
@@ -4,7 +4,7 @@ var ReactDOM = require('react-dom');
 var Router = require('react-router');
 var Link = Router.Link;
 
-var withTeachAPI = require('../mixins/teach-api-client.jsx');
+var withTeachAPI = require('../hoc/with-teach-api.jsx');
 
 var HeroUnit = require('../components/hero-unit.jsx');
 var Map = require('../components/map.jsx');
