+++ test/browser/login.test.jsx
@@ -1,3 +1,4 @@
+var _ = require('lodash');
 var EventEmitter = require('events').EventEmitter;
 var urlParse = require('url').parse;
 var should = require('should');
@@ -6,8 +7,8 @@ var ReactDOM = require('react-dom');
 var TestUtils = require('react-addons-test-utils');
 
 var Login = require('../../components/login.jsx');
-var LoginLink = Login.LoginLink;
-var LogoutLink = Login.LogoutLink;
+var LoginLink = require('../../components/login/LoginLink.jsx');
+var LogoutLink = require('../../components/login/LogoutLink.jsx');
 
 var stubContext = require('./stub-context.jsx');
 var StubTeachAPI = require('./stub-teach-api');
@@ -16,13 +17,14 @@ var StubRouter = require('./stub-router');
 var ADMIN_RE = /administration/i;
 
 describe("Login", function() {
-  var login, teachAPI;
+  var withTeach, login, teachAPI;
 
   beforeEach(function() {
     teachAPI = new StubTeachAPI();
-    login = stubContext.render(Login, {
+    withTeach = stubContext.render(Login, {
       teachAPI: teachAPI
     });
+    login = withTeach.getComponent();
   });
 
   afterEach(function() {
@@ -32,7 +34,12 @@ describe("Login", function() {
   });
 
   it("removes teach API event listeners when unmounted", function() {
-    var events = Object.keys(Login.teachAPIEvents);
+    var events = [
+      'login:start',
+      'login:error',
+      'login:success',
+      'logout'
+    ];
 
     events.forEach(function(event) {
       EventEmitter.listenerCount(teachAPI, event).should.equal(1);
@@ -104,7 +111,10 @@ function renderLink(linkClass, props) {
 
 describe("Login.LoginLink", function() {
   it("should create a link w/ expected callback", function() {
-    var link = renderLink(LoginLink, {origin: 'http://teach'});
+    var link = renderLink(LoginLink, {
+      origin: 'http://teach',
+      loginBaseURL: 'http://teach-api'
+    });
     var info = urlParse(ReactDOM.findDOMNode(link).href, true);
 
     info.protocol.should.eql('http:');
@@ -115,7 +125,10 @@ describe("Login.LoginLink", function() {
   });
 
   it("should accept action='signup'", function() {
-    var link = renderLink(LoginLink, {action: 'signup'});
+    var link = renderLink(LoginLink, {
+      action: 'signup',
+      loginBaseURL: 'http://teach-api'
+    });
     var info = urlParse(ReactDOM.findDOMNode(link).href, true);
 
     info.query.action.should.eql('signup');
@@ -124,7 +137,8 @@ describe("Login.LoginLink", function() {
   it("should accept callbackSearch prop", function() {
     var link = renderLink(LoginLink, {
       origin: 'http://teach',
-      callbackSearch: '?foo=on'
+      callbackSearch: '?foo=on',
+      loginBaseURL: 'http://teach-api'
     });
     var info = urlParse(ReactDOM.findDOMNode(link).href, true);
 
@@ -134,7 +148,10 @@ describe("Login.LoginLink", function() {
 
 describe("Login.LogoutLink", function() {
   it("should create a link w/ expected callback", function() {
-    var link = renderLink(LogoutLink, {origin: 'http://teach'});
+    var link = renderLink(LogoutLink, {
+      origin: 'http://teach',
+      loginBaseURL: 'http://teach-api'
+    });
     var info = urlParse(ReactDOM.findDOMNode(link).href, true);
 
     info.protocol.should.eql('http:');
