+++ test/browser/modal-clubs.test.js
@@ -4,10 +4,16 @@ var should = require('should');
 var React = require('react');
 var ReactDOM = require('react-dom');
 var TestUtils = require('react-addons-test-utils');
-var stubContext = require('./stub-context.jsx');
 var _ = require('underscore');
 
-var ModalAddOrChangeYourClub = require('../../components/modal-clubs.jsx');
+
+var ModalClubs = require('../../components/modal-clubs.jsx');
+
+console.log(ModalClubs);
+
+var TeachAPI = require('./stub-teach-api');
+var stubContext = require('./stub-context.jsx');
+
 var Util = require('../util.js');
 
 var MODAL_ERROR_REGEX = /an error occurred/i;
@@ -30,7 +36,7 @@ function ensureFormFieldsDisabledValue(component, isDisabled) {
   }
 }
 
-describe("ModalClubsGeneral", function() {
+describe("ModalClubs", function() {
   var modal, teachAPI, onSuccess;
 
   beforeEach(function() {
@@ -45,11 +51,14 @@ describe("ModalClubsGeneral", function() {
   });
 
   describe("adding a club", function() {
+
     beforeEach(function() {
-      modal = stubContext.render(ModalAddOrChangeYourClub, {
-        onSuccess: onSuccess
+      teachAPI = new TeachAPI();
+      var withTeach = stubContext.render(ModalClubs, {
+        onSuccess: onSuccess,
+        teachAPI: teachAPI
       });
-      teachAPI = modal.getTeachAPI();
+      modal = withTeach.getComponent();
     });
 
     it("renders", function() {
@@ -89,9 +98,9 @@ describe("ModalClubsGeneral", function() {
     it("has valid labels for form elements", function() {
       teachAPI.emit('username:change', 'foo');
 
-      Util.ensureLabelLinkage(modal, 'ModalAddOrChangeYourClub_name');
-      Util.ensureLabelLinkage(modal, 'ModalAddOrChangeYourClub_website');
-      Util.ensureLabelLinkage(modal, 'ModalAddOrChangeYourClub_description');
+      Util.ensureLabelLinkage(modal, 'ModalClubs_name');
+      Util.ensureLabelLinkage(modal, 'ModalClubs_website');
+      Util.ensureLabelLinkage(modal, 'ModalClubs_description');
     });
 
     it("handles location changes containing JSON", function() {
@@ -234,11 +243,13 @@ describe("ModalClubsGeneral", function() {
     };
 
     beforeEach(function() {
-      modal = stubContext.render(ModalAddOrChangeYourClub, {
+      teachAPI = new TeachAPI();
+      var withTeach = stubContext.render(ModalClubs, {
         club: club,
-        onSuccess: onSuccess
+        onSuccess: onSuccess,
+        teachAPI: teachAPI
       });
-      teachAPI = modal.getTeachAPI();
+      modal = withTeach.getComponent();
       teachAPI.emit('username:change', 'foo');
     });
 
@@ -291,8 +302,8 @@ describe("ModalClubsGeneral", function() {
   });
 });
 
-describe("ModalAddOrChangeYourClub.normalizeClub", function() {
-  var normalizeClub = ModalAddOrChangeYourClub.normalizeClub;
+describe("ModalClubs.normalizeClub", function() {
+  var normalizeClub = ModalClubs.getComponent().normalizeClub;
 
   it("prepends http:// to website if needed", function() {
     normalizeClub({website: 'foo'}).website.should.eql('http://foo');
@@ -311,8 +322,8 @@ describe("ModalAddOrChangeYourClub.normalizeClub", function() {
   });
 });
 
-describe("ModalAddOrChangeYourClub.validateClub", function() {
-  var validateClub = ModalAddOrChangeYourClub.validateClub;
+describe("ModalClubs.validateClub", function() {
+  var validateClub = ModalClubs.getComponent().validateClub;
 
   function club(info) {
     return _.extend({
